% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_annotation_list.R
\name{as_annotation_list}
\alias{as_annotation_list}
\title{Convert an annotation data.frame to a named list, where the names of the 
items are the annotated targets and the entries are unique annotations.}
\usage{
as_annotation_list(annotations, termCol, keyCol)
}
\arguments{
\item{annotations}{a data frame containing some annotations}

\item{termCol}{the name of the column that contains the annotation terms 
(e.g. Gene Ontology terms)}

\item{keyCol}{the name of the column that contains the annotated targets 
(e.g. UniProt accessions)}
}
\value{
a named list where the names are found in keyCol and the entries are 
unique items from termCol
}
\description{
Convert an annotation data.frame to a named list, where the names of the 
items are the annotated targets and the entries are unique annotations.
}
\examples{
# read a GOA file 
go <- read_gpa("goa_human.gpa.gz")
# convert to a named annotation list 
ann <- as_annotation_list(go, "GO.ID", "UNIPROT")
}
