% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_enrichment.R
\name{coexpression_enrichment}
\alias{coexpression_enrichment}
\title{Calculate a Z score for enrichment for coexpression}
\usage{
coexpression_enrichment(network, expr, bootstraps = 100, seed = 0,
  colname = "coexpr")
}
\arguments{
\item{network}{a data frame containing interactions between protein pairs
in the first two columns}

\item{expr}{a matrix containing mRNA or protein expression data, with 
genes as columns}

\item{bootstraps}{the number of times to rewire the network in order
to calculate the Z score; defaults to 100}

\item{seed}{the number to seed the rewiring algorithm with; defaults to zero}

\item{colname}{the column name to assign to the shared annotations 
column created in the network data frame; defaults to "coexpr"}
}
\value{
a list with three objects: the median coexpression in the 
observed network; the median coexpression in each randomized network;
and the Z score for the enrichment
}
\description{
Calculate the enrichment among a set of interacting proteins for mRNA or 
protein co-expression, relative to a rewired version of the same network.
}
