% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_breadth.R
\name{filter_breadth}
\alias{filter_breadth}
\title{Filter very broad or very specific GO terms from a GOA file.}
\usage{
filter_breadth(goa, min.breadth = 20, max.breadth = 1000)
}
\arguments{
\item{min.breadth}{minimum size of a GO group to consider}

\item{max.breadth}{maximum size of a GO group to consider}

\item{go}{the GOA file, read with `read_gpa`}
}
\value{
the filtered GPA file
}
\description{
Filter very broad or very specific GO terms from a GOA file.
}
\examples{
# read a GOA file 
go <- read_gpa("goa_human.gpa.gz")
# get GO terms annotated to between 30 and 100 proteins 
bp <- filter_breadth(go, min.breadth = 30, max.breadth = 100)
}
