% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_roots.R
\name{filter_roots}
\alias{filter_roots}
\title{Filter GO terms from a particular ontological category (BP, CC, or MF) from 
a GOA file.}
\usage{
filter_roots(goa, ontology, root = c("BP", "CC", "MF"))
}
\arguments{
\item{ontology}{the GO ontology, as read with the \code{ontologyIndex} 
package}

\item{root}{the ontological category to restrict the GOA file to; one of
"BP", "CC", or "MF"}

\item{go}{the GO annotation data frame}
}
\value{
the filtered GO annotation file
}
\description{
Filter GO terms from a particular ontological category (BP, CC, or MF) from 
a GOA file.
}
