% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_coexpression.R
\name{network_coexpression}
\alias{network_coexpression}
\title{Calculate enrichment for coexpression among interacting proteins}
\usage{
network_coexpression(network, expr, ...)
}
\arguments{
\item{network}{a matrix or data frame, with nodes (genes) in the first 
two columns}

\item{expr}{mRNA or protein expression data for}

\item{...}{further arguments passed directly to the cor function, such as
coefficient (e.g., Pearson or Spearman)}
}
\value{
the difference in medians between interacting and random protein 
pairs
}
\description{
This function calculates the difference between the median coexpression 
of interacting protein pairs within a network, and random pairs from the same 
expression matrix.
}
