% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_shared_annotations.R
\name{network_shared_annotations}
\alias{network_shared_annotations}
\title{Calculate shared annotations between network nodes}
\usage{
network_shared_annotations(network, shared, col_name = "shared")
}
\arguments{
\item{network}{a data.frame with nodes (genes) in the first two columns}

\item{shared}{a matrix containing the number of shared annotations, e.g. 
GO terms, between each pair of proteins (as output by `shared_annotations`)}

\item{col_name}{the name of the new column to create; defaults to "coexpr"}
}
\value{
the input network, with an extra column containing the number of 
shared annotations (e.g., GO terms)
}
\description{
Calculate the number of shared annotations between each pair of genes or 
proteins in a network, and append this value to the input data frame
as an extra column.
}
\examples{
# read network
net <- read.delim("mouse-network.tsv.gz")
# get proteome-wide shared annotations
shared <- shared_annotations(ann)
# calculate number of shared GO terms for each edge the network
net.shared <- network_shared_go(net, shared)
}
