% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_shared_go.R
\name{network_shared_go}
\alias{network_shared_go}
\title{Calculate the number of shared GO terms between each pair of genes or 
proteins in a network.}
\usage{
network_shared_go(network, shared.go, keep.all = F)
}
\arguments{
\item{network}{a data.frame with nodes (genes) in the first two columns}

\item{shared.go}{a matrix containing the number of shared GO terms between 
each pair of proteins (as output by `proteome_shared_go`)}

\item{keep.all}{if false, only results for network edges where both nodes
were found in the shared GO matrix will be reported}
}
\value{
the input network, with an extra column containing the number of 
shared GO terms
}
\description{
Calculate the number of shared GO terms between each pair of genes or 
proteins in a network.
}
\examples{
# read network
net <- read.delim("mouse-network.tsv.gz")
# get proteome-wide shared GO terms
shared <- proteome_shared_go("goa_mouse.gpa.gz", database = org.Mm.eg.db,
   accession = "ENSEMBL")
# calculate number of shared GO terms for each edge the network
net.shared <- network_shared_go(net, shared)
}
