% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_shared_pairwise_annotations.R
\name{network_shared_pairwise_annotations}
\alias{network_shared_pairwise_annotations}
\title{Calculate shared pairwise annotations between network nodes}
\usage{
network_shared_pairwise_annotations(network, annotations, annotation_pairs,
  col_name = "shared_pairs")
}
\arguments{
\item{network}{a data.frame with nodes (genes) in the first two columns}

\item{annotations}{an annotation list, created by 
\code{\link{as_annotation_list}}, with nodes as names and annotations as 
entries in the list; for example, a list associating proteins with their
Pfam domains}

\item{annotation_pairs}{a logical or binary matrix containing pairs of 
annotations that are considered to be shared; for example, a matrix of
Pfam domains, where known domain-domain interactions are marked with
ones}

\item{col_name}{the name of the new column to create; defaults to 
"shared_pairs"}
}
\value{
the input network, with an extra column containing the number of 
shared pairwise annotations (e.g., domain-domain interactions)
}
\description{
Calculate the number of shared pairwise annotations between each pair of 
genes or proteins in a network, and append this value to the input data frame
as an extra column. By "paired," it is meant that the annotations apply
to pairs of proteins: for instance, interactions between Pfam domains 
derived from analysis of crystal structures (as curated by 3did and other
databases).
}
