% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteome_shared_go.R
\name{proteome_shared_go}
\alias{proteome_shared_go}
\title{Rapidly identify shared GO terms for all pairs of proteins in a proteome.}
\usage{
proteome_shared_go(gpa.file, database = NULL, accession = "UNIPROT",
  filter.NOT = T, filter.evidence = c("ND", "IPI", "IEA", "NAS"),
  filter.breadth = T, min.breadth = 20, max.breadth = 1000,
  ontology.file = NULL, ontology = c("BP", "CC", "MF"))
}
\arguments{
\item{database}{optionally, provide a Bioconductor `AnnotationDbi` database
to map identifiers from UniProt to some other accession}

\item{accession}{optionally, the type of identifier to map UniProt accessions
to using a Bioconductor `AnnotationDbi` database. Must be a keytype in the
database.}

\item{filter.breadth}{if true, filter annotations that are very specific or 
very broad}

\item{min.breadth}{minimum size of a GO group to consider}

\item{max.breadth}{maximum size of a GO group to consider}

\item{ontology.file}{optionally, provide a GO ontology file in order to 
restrict analysis to a specific ontological category (BP, CC, or MF)}

\item{ontology}{if an ontology file is specified, the GO root to consider 
(BP, CC, or MF)}

\item{filepath}{the location of the GPA file. Files can be downloaded from
ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/}

\item{filter_NOT}{if true, filter annotations with the qualifier NOT}

\item{filter_evidence}{optionally, specify evidence codes to filter. By 
default, evidence codes ND, IPI, IEA and NAS are filtered.}
}
\value{
a `dsCMatrix` containing all
}
\description{
Rapidly identify shared GO terms for all pairs of proteins in a proteome.
}
\examples{
# get all shared GO terms between mouse Ensembl genes, considering only 
# BP groups of size 20-100 proteins
shared <- proteome_shared_go("goa_mouse.gpa.gz", database = org.Mm.eg.db,
   accession = "ENSEMBL", filter.breadth = T, max.breadth = 100),
   ontology.file = "go-basic.obo.gz", root = "BP")
}
