% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gaf.R
\name{read_gaf}
\alias{read_gaf}
\title{Read a GAF file}
\usage{
read_gaf(filepath, database = NULL, accession = "UNIPROT", filter.NOT = T,
  filter.evidence = c("ND", "IPI", "IEA", "NAS"), ontology = NULL,
  propagate = T)
}
\arguments{
\item{filepath}{the location of the GAF file. Files can be downloaded from
ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/}

\item{database}{optionally, provide a Bioconductor `AnnotationDbi` database
to map identifiers from UniProt to some other accession}

\item{accession}{optionally, the type of identifier to map UniProt accessions
to using a Bioconductor `AnnotationDbi` database. Must be a keytype in the
database.}

\item{propagate}{if true, and an ontology file is provided, all ancestors of 
a given term are associated with each protein.}

\item{filter_NOT}{if true, filter annotations with the qualifier NOT}

\item{filter_evidence}{optionally, specify evidence codes to filter. By 
default, evidence codes ND, IPI, IEA and NAS are filtered.}
}
\value{
a data.frame containing the filtered GAF file
}
\description{
Read a GO annotation file in GAF format. Further information about the GAF
format is available from the 
\href{http://geneontology.org/page/go-annotation-file-gaf-format-21}{
Gene Ontology Consortium}.
}
\examples{
# read human GOA and map to Ensembl, filtering IEA annotations
go <- read_gaf("goa_human.gpa.gz", database = org.Hs.eg.db,
   accession = "ENSEMBL", filter.evidence = "IEA")
# read mouse GOA and map to gene symbol, filtering IEA and IPI annotations
go <- read_gaf("goa_mouse.gpa.gz", database = org.Mm.eg.db,
   accession = "SYMBOL", filter.evidence = c("IEA", "IPI"))
}
