% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpa.R
\name{read_gpa}
\alias{read_gpa}
\title{Read a GO annotation file in GPA format.}
\usage{
read_gpa(filepath, database = NULL, accession = "UNIPROT", filter.NOT = T,
  filter.evidence = c("ND", "IPI", "IEA", "NAS"), ontology.file = NULL,
  propagate = T)
}
\arguments{
\item{filepath}{the location of the GPA file. Files can be downloaded from
ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/}

\item{database}{optionally, provide a Bioconductor `AnnotationDbi` database
to map identifiers from UniProt to some other accession}

\item{accession}{optionally, the type of identifier to map UniProt accessions
to using a Bioconductor `AnnotationDbi` database. Must be a keytype in the
database.}

\item{filter_NOT}{if true, filter annotations with the qualifier NOT}

\item{filter_evidence}{optionally, specify evidence codes to filter. By 
default, evidence codes ND, IPI, IEA and NAS are filtered.}

\item{propatate}{if true, and an ontology file is provided, all ancestors of 
a given term are associated with each protein.}
}
\value{
a data.frame containing the filtered GPA file
}
\description{
Read a GO annotation file in GPA format.
}
\examples{
# read human GOA and map to Ensembl, filtering IEA annotations
go <- read_gpa("goa_human.gpa.gz", database = org.Hs.eg.db,
   accession = "ENSEMBL", filter.evidence = "IEA")
# read mouse GOA and map to gene symbol, filtering IEA and IPI annotations
go <- read_gpa("goa_mouse.gpa.gz", database = org.Mm.eg.db,
   accession = "SYMBOL", filter.evidence = c("IEA", "IPI"))
}
