% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_annotation_enrichment.R
\name{shared_annotation_enrichment}
\alias{shared_annotation_enrichment}
\title{Calculate a Z score for enrichment for shared annotations}
\usage{
shared_annotation_enrichment(network, shared, bootstraps = 100, seed = 0,
  colname = "shared")
}
\arguments{
\item{network}{a data frame containing interactions between protein pairs
in the first two columns}

\item{shared}{a matrix denoting the number of shared annotations between
each protein pair}

\item{bootstraps}{the number of times to rewire the network in order
to calculate the Z score; defaults to 100}

\item{seed}{the number to seed the rewiring algorithm with; defaults to zero}

\item{colname}{the column name to assign to the shared annotations 
column created in the network data frame; defaults to "shared"}
}
\value{
a list with three objects: the proportion of interacting 
protein pairs that share GO terms in the observed network; the proportions
in each randomized network; and the Z score for the enrichment
}
\description{
Calculate the enrichment among a set of interacting proteins for shared
annotations, relative to a rewired version of the same network.
}
