% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_pairwise_annotations.R
\name{shared_pairwise_annotations}
\alias{shared_pairwise_annotations}
\title{Construct a matrix of shared pairwise annotation counts}
\usage{
shared_pairwise_annotations(annotations, annotation_pairs)
}
\arguments{
\item{annotations}{a list of annotations, with the names of the list 
corresponding to nodes in the network and entries corresponding to 
annotations}

\item{annotation_pairs}{a binary or logical matrix}
}
\value{
a matrix containing the total number of shared pairwise annotations
between each unique gene or protein pair
}
\description{
This function is related to \code{\link{shared_annotations}}, but whereas 
that function calculates the total number of annotations shared by each
pair of genes or proteins, this function calculates the total number of 
pairwise annotations shared by each gene or protein pair. These are 
annotations that are only interpretable in a pairwise manner: for instance,
interactions between Pfam domains.
}
