% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_adjacency_matrix.R
\name{to_adjacency_matrix}
\alias{to_adjacency_matrix}
\title{Convert a data frame to an adjacency matrix}
\usage{
to_adjacency_matrix(dat)
}
\arguments{
\item{dat}{a data frame with the objects that are interacting in the 
first two columns}
}
\value{
a square adjacency matrix with all unique values of the first 
two columns of \code{dat} as row and column names
}
\description{
Convert a data frame of pairwise interactions into an adjacency matrix
between all objects.
}
