% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_roots.R
\name{filter_roots}
\alias{filter_roots}
\title{Filter GO terms from a particular ontological category (BP, CC, or MF) from 
a GOA file.}
\usage{
filter_roots(goa, ontology.file, root = c("BP", "CC", "MF"))
}
\arguments{
\item{ontology.file}{location of the GO ontology file}

\item{root}{the ontological category to restrict the GOA file to; one of
"BP", "CC", or "MF"}

\item{go}{the GPA file, read with `read_gpa`}
}
\value{
the filtered GPA file
}
\description{
Filter GO terms from a particular ontological category (BP, CC, or MF) from 
a GOA file.
}
\examples{
# read a GOA file 
go <- read_gpa("goa_human.gpa.gz")
# get BP terms only
bp <- filter_roots(go, ontology.file = "go-basic.obo.gz", root = "BP")
}
