% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_coexpression.R
\name{network_coexpression}
\alias{network_coexpression}
\title{Calculate the coexpression of each pair of genes or proteins in a network}
\usage{
network_coexpression(network, expr, ...)
}
\arguments{
\item{network}{a matrix or data frame, with nodes (genes) in the first 
two columns}

\item{expr}{a matrix containing mRNA or protein expression data, with 
genes as columns}

\item{...}{further arguments passed directly to the cor function, such as
coefficient (e.g., Pearson or Spearman)}
}
\value{
the input network, with an extra column containing the 
coexpression of each interacting gene or protein pair
}
\description{
Calculate the mRNA or protein coexpression of each pair of genes or 
proteins in a network, and append this value to the input data frame
as an extra column.
}
