% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_shared_annotations.R
\name{network_shared_annotations}
\alias{network_shared_annotations}
\title{Calculate the number of shared annotations between each pair of genes or 
proteins in a network.}
\usage{
network_shared_annotations(network, shared, keep.all = F)
}
\arguments{
\item{network}{a data.frame with nodes (genes) in the first two columns}

\item{shared}{a matrix containing the number of shared annotations, e.g. 
GO terms, between each pair of proteins (as output by `shared_annotations`)}

\item{keep.all}{if false, only results for network edges where both nodes
were found in the shared annotations matrix will be reported}
}
\value{
the input network, with an extra column containing the number of 
shared annotations (e.g., GO terms)
}
\description{
Calculate the number of shared annotations between each pair of genes or 
proteins in a network.
}
\examples{
# read network
net <- read.delim("mouse-network.tsv.gz")
# get proteome-wide shared annotations
shared <- shared_annotations(ann)
# calculate number of shared GO terms for each edge the network
net.shared <- network_shared_go(net, shared)
}
